+++ src/devSVG.cpp
@@ -50,16 +50,18 @@ public:
   std::string clipid;  // ID for the clip path
   double clipx0, clipx1, clipy0, clipy1;  // Save the previous clip path to avoid duplication
   bool standalone;
+  bool fix_text_size;
   const std::string file;
   cpp11::list system_aliases;
   cpp11::list user_aliases;
   cpp11::strings ids;
 
-  SVGDesc(SvgStreamPtr stream_, bool standalone_, cpp11::list aliases_, const std::string& file_, cpp11::strings ids_):
+  SVGDesc(SvgStreamPtr stream_, bool standalone_, cpp11::list aliases_, const std::string& file_, cpp11::strings ids_, bool fix_text_size_):
       stream(stream_),
       pageno(0),
       clipx0(0), clipx1(0), clipy0(0), clipy1(0),
       standalone(standalone_),
+      fix_text_size(fix_text_size_),
       file(file_),
       system_aliases(cpp11::as_cpp<cpp11::list>(aliases_["system"])),
       user_aliases(cpp11::as_cpp<cpp11::list>(aliases_["user"])),
@@ -749,15 +751,16 @@ void svg_text(double x, double y, const char *str, double rot,
   SVGDesc *svgd = (SVGDesc*) dd->deviceSpecific;
   SvgStreamPtr stream = svgd->stream;
 
+  double width = svg_strwidth(str, gc, dd);
   int adj = 0;
   double rad = rot * 2 * M_PI / 360;
   if (hadj == 0.5) {
-    double offset = svg_strwidth(str, gc, dd) / 2;
+    double offset = width / 2;
     adj = 1;
     x -= offset * cos(-rad);
     y -= offset * sin(-rad);
   } else if (hadj == 1) {
-    double offset = svg_strwidth(str, gc, dd);
+    double offset = width;
     adj = 2;
     x -= offset * cos(-rad);
     y -= offset * sin(-rad);
@@ -783,18 +786,23 @@ void svg_text(double x, double y, const char *str, double rot,
     write_style_str(stream, "font-style", "italic");
   if (!is_black(gc->col))
     write_style_col(stream, "fill", gc->col);
-  if (adj == 1)
+  if (adj == 1) {
     write_style_str(stream, "text-align", "center");
-  if (adj == 2)
+    write_style_str(stream, "text-anchor", "middle");
+  }
+  if (adj == 2) {
     write_style_str(stream, "text-align", "right");
+    write_style_str(stream, "text-anchor", "end");
+  }
 
   std::string font = fontname(gc->fontfamily, gc->fontface, svgd->system_aliases, svgd->user_aliases);
   write_style_str(stream, "font-family", font.c_str());
   write_style_end(stream);
 
-  double width = svg_strwidth(str, gc, dd);
-  (*stream) << " textLength='" << width << "px'";
-  (*stream) << " lengthAdjust='spacingAndGlyphs'";
+  if (svgd->fix_text_size) {
+    (*stream) << " textLength='" << width << "px'";
+    (*stream) << " lengthAdjust='spacingAndGlyphs'";
+  }
   stream->put('>');
 
   write_escaped(stream, str);
@@ -869,7 +877,8 @@ void svg_release_mask(SEXP ref, pDevDesc dd) {}
 pDevDesc svg_driver_new(SvgStreamPtr stream, int bg, double width,
                         double height, double pointsize,
                         bool standalone, cpp11::list& aliases,
-                        const std::string& file, cpp11::strings id) {
+                        const std::string& file, cpp11::strings id,
+                        bool fix_text_size) {
 
   pDevDesc dd = (DevDesc*) calloc(1, sizeof(DevDesc));
   if (dd == NULL)
@@ -946,13 +955,13 @@ pDevDesc svg_driver_new(SvgStreamPtr stream, int bg, double width,
   dd->deviceVersion = R_GE_definitions;
 #endif
 
-  dd->deviceSpecific = new SVGDesc(stream, standalone, aliases, file, id);
+  dd->deviceSpecific = new SVGDesc(stream, standalone, aliases, file, id, fix_text_size);
   return dd;
 }
 
 void makeDevice(SvgStreamPtr stream, std::string bg_, double width, double height,
                 double pointsize, bool standalone, cpp11::list& aliases,
-                const std::string& file, cpp11::strings id) {
+                const std::string& file, cpp11::strings id, bool fix_text_size) {
 
   int bg = R_GE_str2col(bg_.c_str());
 
@@ -960,7 +969,7 @@ void makeDevice(SvgStreamPtr stream, std::string bg_, double width, double heigh
   R_CheckDeviceAvailable();
   BEGIN_SUSPEND_INTERRUPTS {
     pDevDesc dev = svg_driver_new(stream, bg, width, height, pointsize,
-                                  standalone, aliases, file, id);
+                                  standalone, aliases, file, id, fix_text_size);
     if (dev == NULL)
       cpp11::stop("Failed to start SVG device");
 
@@ -974,10 +983,10 @@ void makeDevice(SvgStreamPtr stream, std::string bg_, double width, double heigh
 [[cpp11::register]]
 bool svglite_(std::string file, std::string bg, double width, double height,
               double pointsize, bool standalone, cpp11::list aliases,
-              cpp11::strings id) {
+              cpp11::strings id, bool fix_text_size) {
 
   SvgStreamPtr stream(new SvgStreamFile(file, 1));
-  makeDevice(stream, bg, width, height, pointsize, standalone, aliases, file, id);
+  makeDevice(stream, bg, width, height, pointsize, standalone, aliases, file, id, fix_text_size);
 
   return true;
 }
@@ -986,10 +995,10 @@ bool svglite_(std::string file, std::string bg, double width, double height,
 cpp11::external_pointer<std::stringstream> svgstring_(cpp11::environment env, std::string bg,
                                          double width, double height, double pointsize,
                                          bool standalone, cpp11::list aliases,
-                                         cpp11::strings id) {
+                                         cpp11::strings id, bool fix_text_size) {
 
   SvgStreamPtr stream(new SvgStreamString(env));
-  makeDevice(stream, bg, width, height, pointsize, standalone, aliases, "", id);
+  makeDevice(stream, bg, width, height, pointsize, standalone, aliases, "", id, fix_text_size);
 
   SvgStreamString* strstream = static_cast<SvgStreamString*>(stream.get());